Add a gpg key-signing subcommand
This takes the best things about the commands in ~/.stripe and
~/stripe/henson/bin and unifies them, and puts them under a yoyo
umbrella.

The key material (even though it's encrypted) never gets downloaded onto
the local machine, instead gets dumped into a per-key ramdisk.

This can be the basis for the new generate_credentials flow where gpg
key-signing happens outside that whole flow.

The current credentialing process will continue to use the .stripe-based
gpg methods, but we can start phasing that out as soon as we get a
decredentialing command.