Clean out any puppet certificate state before starting marionette
This should ensure that any machines we re-use don't require a manual
cleanup of existing puppet certificates. This subly (but IMO
acceptably and for the better) shifts the trust we put in the machine:

Where previously we would assume that machines' puppet identities
never change, we now accept that they do, and we trust the machine
that we already have established trust with (via the ssh key exchange)
that they are the source of the correct cert.

This should still, as before, catch the case where somebody generates
a certificate for a known serial number: Unless the fingerprint
matches that on the new machine, they won't get their puppet identity
signed. (instead, we'll throw it away automatically!)