Correctly recognize unrevokable keys in gpg_revoke
This is a nasty logic error: We'd skip the global-revocation step if
the key *had* a designated revoker, not if it *didn't have* one.

This is fairly peaceful for the keys we've (only partially) revoked so
far (as we revoke them another way), but it's kinda annoying on keys
that can't be revoked, so let's fix this for greater automation.